feat(dashboard): extract highlightText function for keyword highlighting
